requirements-light.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Task Queue
celery==5.3.4
redis==5.0.1

# HTTP Client
requests==2.31.0

# Database
supabase==2.0.2
psycopg2-binary==2.9.9

# NLP and ML (CPU-only versions)
spacy==3.7.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.35.2
# torch==2.1.1  # Will install CPU version separately

# Data processing
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2

# Web scraping
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
httpx==0.24.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1 